+++ camlib.py
@@ -29,6 +29,7 @@ from rtree import index as rtindex
 from lxml import etree as ET
 from io import StringIO
 import ezdxf
+from ezdxf import recover
 
 import math
 
@@ -1324,7 +1325,11 @@ class Geometry(object):
         self.multigeo = True
 
         # Parse into list of shapely objects
-        dxf = ezdxf.readfile(filename)
+        try:
+            dxf = ezdxf.readfile(filename)
+        except (KeyError, Exception) as e:
+            self.app.log.warning("Standard DXF read failed (%s), trying recovery mode..." % str(e))
+            dxf, auditor = recover.readfile(filename)
         geos = getdxfgeo(dxf, text_mode=text_mode)
 
         # trying to optimize the resulting geometry by merging contiguous lines