+++ defaults.py
@@ -90,7 +90,7 @@ class AppDefaults:
         "global_project_at_startup": False,
         "global_version_check": True,
         "global_send_stats": True,
-        "global_worker_number": int((os.cpu_count()) / 2) if os.cpu_count() > 4 else 1,
+        "global_worker_number": min(4, max(2, int((os.cpu_count()) / 2))) if os.cpu_count() else 2,
         "global_process_number": int((os.cpu_count()) / 4) if os.cpu_count() > 4 else 1,
         "global_tolerance": 0.005,